New commit

+++ 0-bubble_sort.c
@@ -7,11 +7,11 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	size_t i, j, temp;
+	int i, j, temp;
 
-	for (i = 0; i < size - 1; i++)
+	for (i = 0; i < (int) size - 1; i++)
 	{
-		for (j = 0; j < size - i - 1; j++)
+		for (j = 0; j < (int) size - i - 1; j++)
 		{
 			if (array[j] > array[j + 1])
 			{
